requirements.txt
alpaca-trade-api==1.2.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
msgpack==1.0.2
numpy==1.20.3
oauthlib==3.1.1
packaging==20.9
pandas==1.2.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.5
websocket-client==1.1.0
websockets==9.1
google-cloud-storage==1.41.0
